get changes

+++ h2/src/test/org/h2/test/store/TestMVStore.java
@@ -1,6 +1,6 @@
 /*
- * Copyright 2004-2014 H2 Group. Multiple-Licensed under the MPL 2.0,
- * and the EPL 1.0 (http://h2database.com/html/license.html).
+ * Copyright 2004-2019 H2 Group. Multiple-Licensed under the MPL 2.0,
+ * and the EPL 1.0 (https://h2database.com/html/license.html).
  * Initial Developer: H2 Group
  */
 package org.h2.test.store;
@@ -8,9 +8,11 @@ package org.h2.test.store;
 import java.lang.Thread.UncaughtExceptionHandler;
 import java.nio.ByteBuffer;
 import java.nio.channels.FileChannel;
+import java.nio.charset.StandardCharsets;
 import java.util.Iterator;
 import java.util.Map;
 import java.util.Map.Entry;
+import java.util.NoSuchElementException;
 import java.util.Random;
 import java.util.TreeMap;
 import java.util.concurrent.TimeUnit;
@@ -30,6 +32,7 @@ import org.h2.store.fs.FilePath;
 import org.h2.store.fs.FileUtils;
 import org.h2.test.TestBase;
 import org.h2.test.utils.AssertThrows;
+import org.h2.util.Utils;
 
 /**
  * Tests the MVStore.
@@ -121,8 +124,7 @@ public class TestMVStore extends TestBase {
         store.rollback();
         assertTrue(store.hasMap("test"));
         map = store.openMap("test");
-        // TODO the data should get back alive
-        assertNull(map.get("1"));
+        assertEquals("Hello", map.get("1"));
         store.close();
 
         String fileName = getBaseDir() + "/" + getTestName();
@@ -282,7 +284,7 @@ public class TestMVStore extends TestBase {
         map.put(1, new byte[10 * 1024]);
         s.commit();
         MVMap<String, String> meta = s.getMetaMap();
-        Chunk c = Chunk.fromString(meta.get("chunk.1"));
+        Chunk c = Chunk.fromString(meta.get(DataUtils.META_CHUNK+"1"));
         assertTrue(c.maxLen < Integer.MAX_VALUE);
         assertTrue(c.maxLenLive < Integer.MAX_VALUE);
         s.close();
@@ -314,12 +316,13 @@ public class TestMVStore extends TestBase {
 
     private void testVersionsToKeep() throws Exception {
         MVStore s = new MVStore.Builder().open();
+        assertEquals(5, s.getVersionsToKeep());
         MVMap<Integer, Integer> map;
         map = s.openMap("data");
         for (int i = 0; i < 20; i++) {
-            long version = s.getCurrentVersion();
             map.put(i, i);
             s.commit();
+            long version = s.getCurrentVersion();
             if (version >= 6) {
                 map.openVersion(version - 5);
                 try {
@@ -492,9 +495,9 @@ public class TestMVStore extends TestBase {
         }
         long sizeOld = s.getFileStore().size();
         s.compactMoveChunks();
+        s.close();
         long sizeNew = s.getFileStore().size();
         assertTrue("old: " + sizeOld + " new: " + sizeNew, sizeNew < sizeOld);
-        s.close();
     }
 
     private void testBackgroundExceptionListener() throws Exception {
@@ -527,7 +530,7 @@ public class TestMVStore extends TestBase {
                 sleep(10);
             }
             Throwable e = exRef.get();
-            assertTrue(e != null);
+            assertNotNull(e);
             assertEquals(DataUtils.ERROR_WRITING_FAILED,
                     DataUtils.getErrorCode(e.getMessage()));
         } catch (IllegalStateException e) {
@@ -804,7 +807,7 @@ public class TestMVStore extends TestBase {
         String fileName = getBaseDir() + "/" + getTestName();
         FileUtils.delete(fileName);
         MVStore s = openStore(fileName);
-        SequenceMap seq = s.openMap("data", new SequenceMap.Builder());
+        Map<Long,Long> seq = s.openMap("data", new SequenceMap.Builder());
         StringBuilder buff = new StringBuilder();
         for (long x : seq.keySet()) {
             buff.append(x).append(';');
@@ -814,6 +817,9 @@ public class TestMVStore extends TestBase {
     }
 
     private void testCacheSize() {
+        if (config.memory) {
+            return;
+        }
         String fileName = getBaseDir() + "/" + getTestName();
         FileUtils.delete(fileName);
         MVStore s;
@@ -822,6 +828,7 @@ public class TestMVStore extends TestBase {
                 fileName(fileName).
                 autoCommitDisabled().
                 compress().open();
+        s.setReuseSpace(false); // disable free space scanning
         map = s.openMap("test");
         // add 10 MB of data
         for (int i = 0; i < 1024; i++) {
@@ -829,12 +836,15 @@ public class TestMVStore extends TestBase {
         }
         s.close();
         int[] expectedReadsForCacheSize = {
-                3407, 2590, 1924, 1440, 1330, 956, 918
+                1880, 490, 476, 501, 476, 476, 541   // compressed
+//                1887, 1775, 1599, 1355, 1035, 732, 507    // uncompressed
         };
-        for (int cacheSize = 0; cacheSize <= 6; cacheSize += 4) {
+        for (int cacheSize = 0; cacheSize <= 6; cacheSize += 1) {
             int cacheMB = 1 + 3 * cacheSize;
+            Utils.collectGarbage();
             s = new MVStore.Builder().
                     fileName(fileName).
+                    autoCommitDisabled().
                     cacheSize(cacheMB).open();
             assertEquals(cacheMB, s.getCacheSize());
             map = s.openMap("test");
@@ -846,8 +856,14 @@ public class TestMVStore extends TestBase {
             }
             long readCount = s.getFileStore().getReadCount();
             int expected = expectedReadsForCacheSize[cacheSize];
-            assertTrue("reads: " + readCount + " expected: " + expected,
-                    Math.abs(100 - (100 * expected / readCount)) < 5);
+            assertTrue("Cache "+cacheMB+"Mb, reads: " + readCount + " expected: " + expected +
+                    " size: " + s.getFileStore().getReadBytes() +
+                    " cache used: " + s.getCacheSizeUsed() +
+                    " cache hits: " + s.getCache().getHits() +
+                    " cache misses: " + s.getCache().getMisses() +
+                    " cache requests: " + (s.getCache().getHits() + s.getCache().getMisses()) +
+                    "",
+                    Math.abs(100 - (100 * expected / readCount)) < 15);
             s.close();
         }
 
@@ -893,7 +909,7 @@ public class TestMVStore extends TestBase {
         s.close();
         s = openStore(fileName);
         Object test = s.getStoreHeader().get("test");
-        assertFalse(test == null);
+        assertNotNull(test);
         assertEquals("123", test.toString());
         s.close();
     }
@@ -970,14 +986,19 @@ public class TestMVStore extends TestBase {
             }
             ByteBuffer buff = ByteBuffer.allocate(4 * 1024);
             fc.read(buff, i);
-            String h = new String(buff.array(), "UTF-8").trim();
+            String h = new String(buff.array(), StandardCharsets.UTF_8).trim();
             int idx = h.indexOf("fletcher:");
             int old = Character.digit(h.charAt(idx + "fletcher:".length()), 16);
             int bad = (old + 1) & 15;
             buff.put(idx + "fletcher:".length(),
                     (byte) Character.forDigit(bad, 16));
-            buff.rewind();
-            fc.write(buff, i);
+
+            // now intentionally corrupt first or both headers
+            // note that headers may be overwritten upon successfull opening
+            for (int b = 0; b <= i; b += blockSize) {
+                buff.rewind();
+                fc.write(buff, b);
+            }
             fc.close();
 
             if (i == 0) {
@@ -1319,6 +1340,7 @@ public class TestMVStore extends TestBase {
         MVMap<String, String> m = s.openMap("data");
         s.commit();
         long first = s.getCurrentVersion();
+        assertEquals(1, first);
         m.put("0", "test");
         s.commit();
         m.put("1", "Hello");
@@ -1332,7 +1354,9 @@ public class TestMVStore extends TestBase {
         m.put("2", "Welt");
         MVMap<String, String> mFirst;
         mFirst = m.openVersion(first);
-        assertEquals(0, mFirst.size());
+        // openVersion() should restore map at last known state of the version specified
+        // not at the first known state, as it was before
+        assertEquals(1, mFirst.size());
         MVMap<String, String> mOld;
         assertEquals("Hallo", m.get("1"));
         assertEquals("Welt", m.get("2"));
@@ -1340,8 +1364,9 @@ public class TestMVStore extends TestBase {
         assertEquals("Hello", mOld.get("1"));
         assertEquals("World", mOld.get("2"));
         assertTrue(mOld.isReadOnly());
-        s.getCurrentVersion();
-        long old3 = s.commit();
+        long old3 = s.getCurrentVersion();
+        assertEquals(3, old3);
+        s.commit();
 
         // the old version is still available
         assertEquals("Hello", mOld.get("1"));
@@ -1360,9 +1385,16 @@ public class TestMVStore extends TestBase {
         assertEquals("Hi", m.get("1"));
         assertEquals(null, m.get("2"));
 
+        // This test tries to cast in bronze some peculiar behaviour,
+        // which is rather implementation artifact then intentional.
+        // Once store is closed, only one single version of the data
+        // will exists upon re-opening - the latest.
+        // I hope nobody relies on this "multi-versioning".
+/*
         mOld = m.openVersion(old3);
         assertEquals("Hallo", mOld.get("1"));
         assertEquals("Welt", mOld.get("2"));
+*/
 
         try {
             m.openVersion(-3);
@@ -1399,6 +1431,8 @@ public class TestMVStore extends TestBase {
             }
         }
         assertTrue(s.compact(100, 50 * 1024));
+        // compaction alone will not guarantee file size reduction
+        s.compactMoveChunks();
         s.close();
         long len2 = FileUtils.size(fileName);
         assertTrue("len2: " + len2 + " len: " + len, len2 < len);
@@ -1416,9 +1450,8 @@ public class TestMVStore extends TestBase {
             assertEquals(i + 1, m.size());
         }
         assertEquals(1000, m.size());
-        // previously (131896) we fail to account for initial root page for every map
-        // there are two of them here (meta and "data"), hence lack of 256 bytes
-        assertEquals(132152, s.getUnsavedMemory());
+        // memory calculations were adjusted, so as this out-of-the-thin-air number
+        assertEquals(93635, s.getUnsavedMemory());
         s.commit();
         assertEquals(2, s.getFileStore().getWriteCount());
         s.close();
@@ -1429,7 +1462,7 @@ public class TestMVStore extends TestBase {
         assertEquals(0, m.size());
         s.commit();
         // ensure only nodes are read, but not leaves
-        assertEquals(45, s.getFileStore().getReadCount());
+        assertEquals(5, s.getFileStore().getReadCount());
         assertTrue(s.getFileStore().getWriteCount() < 5);
         s.close();
     }
@@ -1495,9 +1528,10 @@ public class TestMVStore extends TestBase {
         assertTrue(s.hasUnsavedChanges());
         s.rollbackTo(v2);
         assertFalse(s.hasUnsavedChanges());
-        assertNull(meta.get("name.data1"));
+        assertNull(meta.get(DataUtils.META_NAME+"data1"));
         assertNull(m0.get("1"));
         assertEquals("Hello", m.get("1"));
+        // no changes - no real commit here
         assertEquals(2, s.commit());
         s.close();
 
@@ -1505,9 +1539,9 @@ public class TestMVStore extends TestBase {
         s.setRetentionTime(45000);
         assertEquals(2, s.getCurrentVersion());
         meta = s.getMetaMap();
-        assertTrue(meta.get("name.data") != null);
-        assertTrue(meta.get("name.data0") != null);
-        assertNull(meta.get("name.data1"));
+        assertNotNull(meta.get(DataUtils.META_NAME + "data"));
+        assertNotNull(meta.get(DataUtils.META_NAME + "data0"));
+        assertNull(meta.get(DataUtils.META_NAME + "data1"));
         m = s.openMap("data");
         m0 = s.openMap("data0");
         assertNull(m0.get("1"));
@@ -1600,13 +1634,13 @@ public class TestMVStore extends TestBase {
         assertNull(s.getMapName(s.getMetaMap().getId()));
         assertNull(s.getMapName(data.getId() + 1));
 
-        String id = s.getMetaMap().get("name.data");
-        assertEquals("name:data", m.get("map." + id));
+        String id = s.getMetaMap().get(DataUtils.META_NAME + "data");
+        assertEquals("name:data", m.get(DataUtils.META_MAP + id));
         assertEquals("Hello", data.put("1", "Hallo"));
         s.commit();
-        assertEquals("name:data", m.get("map." + id));
+        assertEquals("name:data", m.get(DataUtils.META_MAP + id));
         assertTrue(m.get("root.1").length() > 0);
-        assertTrue(m.containsKey("chunk.1"));
+        assertTrue(m.containsKey(DataUtils.META_CHUNK + "1"));
 
         assertEquals(2, s.getCurrentVersion());
 
@@ -1719,6 +1753,7 @@ public class TestMVStore extends TestBase {
         String fileName = getBaseDir() + "/" + getTestName();
         FileUtils.delete(fileName);
         MVStore s = openStore(fileName, 1000);
+        s.setAutoCommitDelay(0);
         MVMap<Integer, String> m = s.openMap("data");
         int factor = 100;
         for (int j = 0; j < 10; j++) {
@@ -1730,24 +1765,15 @@ public class TestMVStore extends TestBase {
         s.close();
 
         s = openStore(fileName);
+        s.setAutoCommitDelay(0);
         s.setRetentionTime(0);
 
         Map<String, String> meta = s.getMetaMap();
-        int chunkCount1 = 0;
-        for (String k : meta.keySet()) {
-            if (k.startsWith("chunk.")) {
-                chunkCount1++;
-            }
-        }
+        int chunkCount1 = getChunkCount(meta);
         s.compact(80, 1);
         s.compact(80, 1);
 
-        int chunkCount2 = 0;
-        for (String k : meta.keySet()) {
-            if (k.startsWith("chunk.")) {
-                chunkCount2++;
-            }
-        }
+        int chunkCount2 = getChunkCount(meta);
         assertTrue(chunkCount2 >= chunkCount1);
 
         m = s.openMap("data");
@@ -1760,12 +1786,7 @@ public class TestMVStore extends TestBase {
         }
         assertFalse(s.compact(50, 1024));
 
-        int chunkCount3 = 0;
-        for (String k : meta.keySet()) {
-            if (k.startsWith("chunk.")) {
-                chunkCount3++;
-            }
-        }
+        int chunkCount3 = getChunkCount(meta);
 
         assertTrue(chunkCount1 + ">" + chunkCount2 + ">" + chunkCount3,
                 chunkCount3 < chunkCount1);
@@ -1776,6 +1797,16 @@ public class TestMVStore extends TestBase {
         s.close();
     }
 
+    private int getChunkCount(Map<String, String> meta) {
+        int chunkCount = 0;
+        for (String k : meta.keySet()) {
+            if (k.startsWith(DataUtils.META_CHUNK)) {
+                chunkCount++;
+            }
+        }
+        return chunkCount;
+    }
+
     private void testCompact() {
         String fileName = getBaseDir() + "/" + getTestName();
         FileUtils.delete(fileName);
@@ -1784,11 +1815,17 @@ public class TestMVStore extends TestBase {
             sleep(2);
             MVStore s = openStore(fileName);
             s.setRetentionTime(0);
+            s.setVersionsToKeep(0);
             MVMap<Integer, String> m = s.openMap("data");
             for (int i = 0; i < 100; i++) {
                 m.put(j + i, "Hello " + j);
             }
-            s.compact(80, 1024);
+            trace("Before - fill rate: " + s.getFillRate() + "%, chunks fill rate: "
+                    + s.getChunksFillRate() + ", len: " + FileUtils.size(fileName));
+            s.compact(80, 2048);
+            s.compactMoveChunks();
+            trace("After  - fill rate: " + s.getFillRate() + "%, chunks fill rate: "
+                    + s.getChunksFillRate() + ", len: " + FileUtils.size(fileName));
             s.close();
             long len = FileUtils.size(fileName);
             // System.out.println("   len:" + len);
@@ -1826,6 +1863,7 @@ public class TestMVStore extends TestBase {
             sleep(2);
             MVStore s = openStore(fileName);
             s.setRetentionTime(0);
+            s.setVersionsToKeep(0);
             MVMap<Integer, String> m = s.openMap("data");
             for (int i = 0; i < 10; i++) {
                 m.put(i, "Hello");
@@ -1841,7 +1879,7 @@ public class TestMVStore extends TestBase {
                 initialLength = len;
             } else {
                 assertTrue("len: " + len + " initial: " + initialLength + " j: " + j,
-                        len <= initialLength * 5);
+                        len <= initialLength * 3);
             }
         }
     }
@@ -1955,7 +1993,7 @@ public class TestMVStore extends TestBase {
             assertEquals(i, it.next().intValue());
         }
         assertFalse(it.hasNext());
-        assertNull(it.next());
+        assertThrows(NoSuchElementException.class, it).next();
         for (int j = 0; j < 10; j++) {
             it = m.keyIterator(j);
             for (int i = j; i < 10; i++) {
